designC/requirements.txt
# Design C: PyTorch VRN with GPU Acceleration
# Python dependencies for end-to-end GPU inference

# Core dependencies (inherit from Design B)
torch>=2.1.0
pytorch3d>=0.7.5
numpy>=1.26.0

# Mesh processing
scikit-image>=0.22.0
trimesh>=4.0.0

# Training & monitoring
tensorboard>=2.15.0
torchsummary>=1.5.1

# Benchmarking and visualization
matplotlib>=3.8.0
seaborn>=0.13.0
pandas>=2.0.0

# Utilities
tqdm>=4.66.0
Pillow>=10.0.0

# Optional: advanced mesh operations
pymeshlab>=2022.2  # for voxelization if needed
